Excluding unix dependant features from build on windows.
